requirements.txt
Flask==2.2.2
google-generativeai==0.1.0
google-auth==2.12.0
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
gunicorn==20.1.0
